% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_for_RDA.R
\name{check_data_for_RDA}
\alias{check_data_for_RDA}
\title{check_data_for_RDA function}
\usage{
check_data_for_RDA(data_quant, initial_data, factor_names)
}
\arguments{
\item{data_quant}{Data frame of numeric values, generally transformed and scalled}

\item{initial_data}{Initial data frame, including both numeric and factor columns}

\item{factor_names}{Character vector of considered factor variables/columns : c("Variable1", "Variable2")}
}
\value{
Outputs are saved in a data frame and conclusions are displayed in console.
}
\description{
This function check if the factor variables used for RDA fit the required conditions.
Results are grouped and conclusions go / no go are displayed in the console and saved in a table.
This function is mainly derived from the explanations of Maxime Herve
\url{https://www.maximeherve.com/r-et-statistiques}
}
\examples{

check_data_for_RDA(mtcars[,1:7], mtcars, c("gear","carb"))

}
