% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{prep_data function}
\usage{
prep_data(data, quantitative_columns, transform_data_method, scale_data)
}
\arguments{
\item{data}{A data frame with at least one numeric column}

\item{quantitative_columns}{Numeric vector of numeric columns position : c(1,2,4)}

\item{transform_data_method}{Method of quantitative data transformation : "log", "sqrt4", "clr", "none"}

\item{scale_data}{TRUE or FALSE. TRUE => scaling of quantitative columns}
}
\value{
The transformed/scaled quantitative data frame
}
\description{
This function helps user to transform and scale quantitative defined data (columns identitified from a given dataset)
for better PCA and RDA outputs.
}
\examples{

prep_data(mtcars, c(1:7), "log", TRUE)


}
